backend/pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = "Store Backend using DRF and Django"
authors = ["Leo-Mahendra <mahender.kurikyala11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
razorpay = "^1.4.1"
gunicorn = "^21.2.0"
cloudinary = "^1.37.0"
django-cloudinary-storage = "^0.3.0"


[tool.poetry.group.dev.dependencies]
asgiref = "3.7.2"
boto3 = "1.28.74"
botocore = "1.31.74"
certifi = "2023.7.22"
cffi = "1.16.0"
charset-normalizer = "3.3.1"
cryptography = "41.0.5"
defusedxml = "0.7.1"
django = "4.2.7"
django-cors-headers = "4.3.0"
django-ses = "3.5.0"
django-templated-mail = "1.1.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.0"
djoser = "2.2.0"
idna = "3.4"
jmespath = "1.0.1"
oauthlib = "3.2.2"
pillow = "10.1.0"
psycopg2 = "2.9.9"
pycparser = "2.21"
pyjwt = "2.8.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python3-openid = "3.2.0"
pytz = "2023.3.post1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
s3transfer = "0.7.0"
six = "1.16.0"
social-auth-app-django = "5.4.0"
social-auth-core = "4.4.2"
sqlparse = "0.4.4"
tzdata = "2023.3"
urllib3 = "2.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
